+++ convertJs/plc-simulator/src/components/index.ts
@@ -1,0 +1,9 @@
+/**
+ * UI Components
+ * Exports centralizados
+ */
+
+export * from './CodeEditor'
+export * from './IOPanel'
+export * from './ControlBar'
+export * from './FileMenu'