+++ convertJs/plc-simulator/src/hooks/index.ts
@@ -4,3 +4,4 @@
  */
 
 export * from './useTimerUpdate'
+export * from './usePLCCycle'